+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 - minor logic change in ui
 - implement initial switch state for game (in gamelist.js)
+- implement game initialisation in gamelist.js
 
 ## 10/05/18